% File TBSSurvival/man/alloyT7987.Rd
% Part of the TBSSurvival package for R (http://www.R-project.org)
% Copyright (C) 2012, Adriano Polpo, Cassio de Campos, Debajyoti Sinha
%                     Jianchang Lin and Stuart Lipsitz.
% Distributed under GPL 3 or later

\name{alloyT7987}

%%%
\alias{alloyT7987}
\title{Data set of...}
\description{
alloyT7987
}
\usage{
TO BE COMPLETED...
}
\arguments{
...
}
\value{
...
}
\seealso{
...
}
\examples{
##...
}
\keyword{TBS}
