% File TBSSurvival/man/tbs.survreg.Rd
% Part of the TBSSurvival package for R (http://www.R-project.org)
% Copyright (C) 2012, Adriano Polpo, Cassio de Campos, Debajyoti Sinha
%                     Jianchang Lin and Stuart Lipsitz.
% Distributed under GPL 3 or later

\name{tbs.survreg}

%%%
\alias{tbs.survreg}

\title{MLE for TBS Failure Model}
\description{
  Do not use this function. See \code{tbs.survreg.mle}.
}
\seealso{
  \code{\link{tbs.survreg.mle}}.
}

